pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apiup"
version = "0.1.0"
description = "Start a local mock REST API server from an OpenAPI 3.x spec. Convention: reads ~/.openapi/spec.yaml by default."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "roebi (Robert Halter)", email = "roebi@users.noreply.github.com" },
]
keywords = ["openapi", "mock", "api", "server", "cli", "rest", "litestar"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: Other/Proprietary License",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Software Development :: Testing :: Mocking",
  "Topic :: Utilities",
]
requires-python = ">=3.11"
dependencies = [
  "litestar[standard]>=2.0",
  "uvicorn>=0.29",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "mypy>=1.10",
  "ruff>=0.4",
  "types-PyYAML",
]

[project.scripts]
apiup = "apiup.cli:main"

[project.urls]
Homepage  = "https://github.com/roebi/apiup"
Repository = "https://github.com/roebi/apiup"
Issues    = "https://github.com/roebi/apiup/issues"
Changelog = "https://github.com/roebi/apiup/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/apiup"]

[tool.hatch.build.targets.sdist]
include = [
  "src/",
  "tests/",
  "README.md",
  "LICENSE",
  "CHANGELOG.md",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
ignore_missing_imports = true
